requirements.txt
# Core ML
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4

# Data validation
great-expectations==0.18.19

# Feature store
duckdb==0.10.3

# Experiment tracking
mlflow==2.13.0

# Pipeline versioning
dvc==3.51.2

# Serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
prometheus-fastapi-instrumentator==7.0.0

# Monitoring
evidently==0.4.30

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pyarrow==15.0.2
requests==2.32.2
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
